+++ packages/components/organisms/f-footer/CHANGELOG.md
@@ -4,6 +4,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+Latest (add to next release)
+------------------------------
+*February 24, 2021*
+
+### Changed
+- Restructured component object into page object model
+- Refactored component and accessibility tests
+
+
 v4.8.3
 ------------------------------
 *February 9, 2021*